pyproject.toml
[tool.poetry]
name = "seems-frontend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9.9"
streamlit = "^1.5.0"
plotly = "^5.5.0"
st-annotated-text = "^3.0.0"
matplotlib = "^3.5.1"
watchdog = "^2.1.6"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
